# cmd/service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a service for refscaler app",
        "title": "RefScaler",
        "contact": {
            "name": "Adrian Grzemski",
            "email": "adrian.grzemski@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/units/{name}": {
            "get": {
                "description": "Returns a specific unit group by its name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "units"
                ],
                "summary": "Get unit by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unit group name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "description": "A list of units representing a group",
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UnitJSON"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.UnitJSON": {
            "description": "A unit with value and available aliases",
            "type": "object",
            "properties": {
                "aliases": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
...
    }
}